refactor: session logic n disabling for CI
